Add support for jumping to commands
Same process as functions: verbose print their details, look for the
definition location, and jump there.

Using a liberal search query to find the command to avoid listing all
permutations of arguments. It's non-greedy and I don't think any
arguments can contain spaces, so it should be good.

I couldn't figure out a more elegant way to fit it into the dispatch
system (which seems to be premised on there being two symmetrical
options), so I just put it at the top.

Include a test.